Updated license
Projects using this language may need it to be closed source. As much as I hate people using my work and not distributing their changes, I do want this to be a language that larger projects can utilize. Plus, still not sure if the finished JIT compiler will utilize the source code or an intermediate language, so until then, I'll assume that users will dump the entire compiler into their games.