Expanded on Module, see comment below.
I've decided that a "pre-compiled" module will just contain that module's AST. Probably not the best way to go about it, but the alternative is taking precompiled SPIR-V and renaming the SSAs. Not fun nor efficient. Also not flexible or compatible with LLVM. Finally, could write a Caliburn Intermediary Representation... but, no. Just, no.