Fixed string insertion logic
Turns out, you don't need to add another int at the end of a copied string unless the length is divisible by 4. If it's not divisible, then there's going to be a null character at the end of the string anyway. I thought you needed a full 32-bit 0 at the end.