CLLR no longer stores ops in shared_ptrs
After a lot of benchmarking, it would seem that, under some circumstances, this produces more consistent performance. Further testing should be done. Reducing the number of benches made the characteristics more apparently. When doing 30+ runs, there's more performance spikes, offset by quicker runs, which made this change more of a wash. But at 20, there is a noteworthy performance improvement. Plus storing ops in a shared_ptr was silly in the first place, but it was a worthwhile change at the time because of... some possible operation which is no longer a thing. So overall, I'm willing to commit this change and call it here.