Renamed float types to fp
So float## is now fp##. This change... Well I've been stewing on it for a while. See, I want to maintain a balance between short and long type names. I feel like going too far in either direction is a bad idea. Float32 felt a little awkward, so I've decided to just change it. Ints will remain as int## and uint##. Is this consistent? Maybe not entirely. But I doubt I'll ever go over to i32 and such. Those type names are too short for my liking.