Adjusted type checker, moved scope impl to .CPP
Type checking now has separated RHS and result args. This means RHS is now immutable, and with that, a new degree of error checking and bad code prevention.

I also intended to make return statements type check, but I forgot I don't have access to the function containing a ScopeStatement, so an alternative has to be found.