Fixed function code generating within functions
The root of the issue comes from Caliburn's lazy function evaluation: function definition is deferred until needed, which prevents unneeded functions from being written to the final shader code. Because of this, if you called a function, that function would get written immediately, which would put the code within another function's body. This would then get translated with no changes to SPIR-V.

So to combat this, this commit adds code sections to the CLLR assembler. Sections are stack-based, meaning they have to be started and ended. When a section is ended, it's written to the final CLLR output. This abstraction is completely hidden to the rest of the code, so only the function and shader stage code has to worry about them.

This commit also fixes some miscellaneous bugs.